Moved routers into their own directories
