+++ server.js
@@ -1,13 +1,13 @@
 const express = require('express');
 
-const mainRouter = require('./api/main-router');
-const accountsRouter = require('./api/account-router');
+const welcomeRouter = require('./api/welcome/welcome-router');
+const accountsRouter = require('./api/accounts/accounts-router');
 
 const server = express();
 
 server.use(express.json());
 
-server.use('/api/', mainRouter);
+server.use('/api/', welcomeRouter);
 server.use('/api/accounts', accountsRouter);
 
 module.exports = server;
